requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
feedparser==6.0.11
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0

# New dependencies for summariser
openai==1.23.3
readability-lxml==0.8.1
jsonschema==4.23.0
lxml[html_clean]
